[PROG-9] Reorganize code to prepare for multiple look ahead agents
