+++ src/pcg/agent_generators/agents/look_ahead/basic_look_ahead_agent.h
@@ -1,0 +1,9 @@
+#pragma once
+#include "src/pcg/agent_generators/agents/look_ahead/look_ahead_agent.h"
+
+class BasicLookAheadAgent : public  LookAheadAgent{
+
+};
+
+
+
