+++ src/pcg/agent_generators/agents/look_ahead/utils/spatial_iterators.h
